Use older python version to fix json encoding with twisted in synapse
Also pin flask to a version that works correctly with both the topologizer and python3.6.
The 2020 resolver allows to keep the rest of the dependencie constraints relaxed as it
is able to better find incompatibilities betwenn packages.
Lastly, the cryptography package is now explicitely specified with a pinned version in Synapse.